Merge do mastera

+++ src/main/java/com/Lernado/managers/LessonRepository.java
@@ -3,4 +3,4 @@ package com.Lernado.managers;
 import com.Lernado.model.Lesson;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LessonRepository extends JpaRepository<Lesson, Long> {}
+public interface LessonRepository extends JpaRepository<Lesson, Integer> {}
