+++ src/main/java/com/Lernado/managers/MaterialRepository.java
@@ -4,4 +4,4 @@ package com.Lernado.managers;
 import com.Lernado.model.Material;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MaterialRepository extends JpaRepository<Material, Long> {}
+public interface MaterialRepository extends JpaRepository<Material, Integer> {}
